UPSTREAM: net: ipa: allow arbitrary number of interconnects
Currently we assume that the IPA hardware has exactly three
interconnects.  But that won't be guaranteed for all platforms,
so allow any number of interconnects to be specified in the
configuration data.

For each platform, define an array of interconnect data entries
(still associated with the IPA clock structure), and record the
number of entries initialized in that array.

Loop over all entries in this array when initializing, enabling,
disabling, or tearing down the set of interconnects.

With this change we no longer need the ipa_interconnect_id
enumerated type, so get rid of it.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit ea151e1915ebef316893f1fdae6c2cd89ae3371b)

BUG=b:175328613
TEST=Modem functional, No IPA hang in suspend stress test

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: I6d06a04a48e3be2f01d3c033c8bc2074aadf122d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2946193
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>