UPSTREAM: media: rockchip: rkisp1: carry ip version information
The IP block evolved from its rk3288/rk3399 base and the vendor
designates them with a numerical version. rk3399 for example
is designated V10 probably meaning V1.0.

There doesn't seem to be an actual version register we could read that
information from, so allow the match_data to carry that information
for future differentiation.

Also carry that information in the hw_revision field of the media-
controller API, so that userspace also has access to that.

The added versions are:
- V10: at least rk3288 + rk3399
- V11: seemingly unused as of now, but probably appeared in some soc
- V12: at least rk3326 + px30
- V13: at least rk1808

[fix checkpatch warning don't use multiple blank lines]

Signed-off-by: Heiko Stuebner <heiko.stuebner@theobroma-systems.com>
Signed-off-by: Dafna Hirschfeld <dafna.hirschfeld@collabora.com>
Reviewed-by: Ezequiel Garcia <ezequiel@collabora.com>
Acked-by: Helen Koike <helen.koike@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit fc672d806bd77eff26117479e90ccdcfd2a8ecb4)

BUG=b:197575417
TEST=Camera works on Scarlet

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: Ic90e31f026fc41089138c1a3e23a79f255e4ec04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3168485
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>