FROMGIT: drm/i915: Add pipe ddb entries into the dbuf state
The dbuf state will be where we collect all the inter-pipe dbuf
allocation stuff. Start by adding the actual per-pipe ddb entries
there.

Originally the plan was to move them there outright, but that no longer
works as we're no longer guaranteed to have a dbuf state when it comes
time to sanity check the ddb overlaps in skl_commit_modeset_enables().
I think when I wrote this originally we did the watermark/ddb
calculation last, and so we couldn't have any crtcs in the state w/o
also having the dbuf state. But that has since changed and we do the
watermark/ddb calculation much earlier, and thus it is now possible
to commit crtcs w/o a dbuf state. So we keep another copy of the
information in the crtc state.

v2: Rebase
v3: Duplicate the entries instead of moving

Reviewed-by: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210122205633.18492-6-ville.syrjala@linux.intel.com
(cherry picked from commit 47a14955106f71e9d6cb2a807bd30dacd403a3fe
https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

BUG=b:179453336
TEST=eDP comes up on brya when built with this commit https://crrev.com/i/3584703

Change-Id: I3db42228f10ce082710a81606bbc7fdb45845e47
Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>