FROMGIT: opp: Defer acquiring the clk until OPPs are added
We acquire the clk at the time the OPP table is allocated, though it
works fine, it is not the best place to do so. One of the main reason
being we may need to acquire it again from dev_pm_opp_set_clkname() if
the platform wants another clock to be acquired instead.

There is also requirement from some of the platforms where they do not
want the OPP core to manage the clock at all.

This patch hence defers acquiring the clk until the time we are certain
about which clk we need to acquire and if we really need to acquire one.
With this commit, the clk will get acquired either from
dev_pm_opp_set_clkname() or while we initialize the OPPs within the
table.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Tested-by: Dmitry Osipenko <digetx@gmail.com>

(cherry picked from commit 32439ac7535a8eddfa016c62ca66ce33b7df1573
https://git.kernel.org/pub/scm/linux/kernel/git/vireshk/pm.git opp/linux-next)

BUG=b:129039370
TEST=boot kukui and check cci devfreq governor

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: I978fd71ff92d626de0b23550e31c0c14c8b2558f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2675527
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>