misc: cs40l25: Careful to not overwrite RX2 slot config
When updating the configuration for the RX1 slot regmap_update_bits
is used such that only the RX1 field is updated. However when this
value is added to the wseq queue, the value of the rest of the
register is ignored. This means that the RX2 configuration is always
set to slot 0. If RX1 is also set to slot 0, which it normally is,
then both RX's are set to the same slot. This causes the hardware
to flag a configuration error which causes A2H to not process audio.

Change-Id: I42317cd2474f5a8fb34d4aebed70ce4ff4443631
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>