ASoC: cs40l25: Update some variable naming
The codec private data pointer is sometimes referred to as codec and sometimes
as cs40l2x. The codec name is confusing with older backports which still
use snd_soc_codec and the cs40l2x naming is confusing with the parent
driver private data. Rename both bases to priv which should be suitable
in all cases. Whilst doing some rename a couple other variable to
increase consistency in the file.

Change-Id: I6021294ea8f6342de00c2b4b8055c35764754bc9
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>