CHROMIUM: iwl7000: fw: fix powersave debugfs out-of-bounds read
The code here checked only that the beginning of the field
(offsetof) fits into the data returned by the firmware, not
that the *entire* field fits. Fix this to use offsetofend()
and see that is <= the size to fix this.

Change-Id: If411638e985242036476192620442cbd179d9a30
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: d527d4cedd67bcab2e0b4d3d7c8c4b72befae90e