ASoC: cs40l25: Remove incorrect usage of params_physical_width
params_physical_width refers to the size of the data in memory and
should not be used for slot size on the bus. If the slot size needs to
change that should be controlled through the ASoC TDM callbacks,
otherwise it makes sense to use the datasize as the slot size.

Change-Id: I312388654021295a603b30b7fda411a67faa4067
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>