ASoC: cs40l25: Don't write DAI format directly from cs40l2x_set_dai_fmt
The DAI format can't be written directly in cs40l2x_set_dai_fmt as this
function is not normally called as part of the audio path bring up it is
called much earlier during initialisation. That means that the device
may go to hibernate between cs40l2x_set_dai_fmt being called and the
audio path being started, causing the register values written to be
lost. Update the code to cache these and write them out during hw_params
when we know the device will be active.

Change-Id: Ia7db3b6783226372b25a1021a77b345efeb9c167
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>