ASoC: cs40l25: Don't write ASP rate directly from set_sysclk
The ASP rate can't be written directly cs40l2x_component_set_sysclk as
this function is not normally called as part of the audio path bring up
it is called much earlier during initialisation. That means that the device
may go to hibernate between cs40l2x_component_set_sysclk being called and the
audio path being started, causing the register values written to be
lost. Update the code to cache these and write them out during hw_params
when we know the device will be active.

Secondly, it makes more sense to set the PLL from the BCLK rate as that
is all we currently support and that is available in hw_params before we
need to change the clocking. However, as removing the set_sysclk
callback would could cause errors in existing machine drivers, leave the
callback and update hardware params to check the two rates match.

Change-Id: Ia46d42f98f8cbf40410151d1132ccc55d36a9ad3
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>