ASoC: cs40l25: Remove local header file
The sound cs40l25 header file basically just contains duplicates of the
information in the devices main header file, however in general the main
header file uses names that are closer to the datasheet. Remove the
local file and update the sound driver to use the main defines. In cases
where the primary define doesn't match the datasheet update that as
well.

Change-Id: Ieef0543a1d07d60f4ff87fe9576e76776e2c5821
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>