ASoC: cs35l45: Store I2C driver max quirk length in cs35l45_private
The I2C driver may have a quirk that defines its max read length.
If data transferred on I2C exceeds the max length, the I2C driver
reports errors and terminates the transfer.

The driver should store the max I2C read length and check the size
before transferring a block of data via I2C. This is most likely to
impact when transferring coeff data from/to the DSP, so convert the
length into number of DSP words and store it.

Change-Id: I5351d4082cb8f0aded1baf98c77e176f312de9de
Signed-off-by: Qi Zhou <qizhou@opensource.cirrus.com>