UPSTREAM: net: ipa: use a Boolean rather than count when replenishing
The count argument to ipa_endpoint_replenish() is only ever 0 or 1,
and always will be (because we always handle each receive buffer in
a single transaction).  Rename the argument to be add_one and change
it to be Boolean.

Update the function description to reflect the current code.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 9af5ccf32383005070092e51b15cee51584323c0)

BUG=b:175328613
TEST=Modem functional, No IPA hang in suspend stress test

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: I995f3ad64a8278f82d08c453d086d2e3df66d7bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2946197
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>