ASoC: cs40l26: Remove incorrect usage of params_physical_width
params_physical_width refers to the size of the data in memory
and should not be used for slot size on the bus. If the slot
size needs to change that should be controlled through the
ASoC TDM callbacks, otherwise it makes sense to use the
datasize as the slot size.

Change-Id: I162e1b1d3812e77b7085c9f1f430ad48e104628c
Signed-off-by: Fred Treven <fred.treven@cirrus.com>