ASoC: cs40l26: Don't write DAI format directly from cs40l26_set_dai_fmt
The DAI format can't be written directly in cs40l26_set_dai_fmt as this
function is not normally called as part of the audio path bring up; it
is called much earlier during initialisation. That means the device may
go go to hibernate between cs40l26_det_dai_fmt being called and the
audio path being started, causing the register calues written to be lost.
Update the code to cache these and write them out during hw_params when
we know the device will be active.

Change-Id: I73bf44093e5e4addbdc8223855acc8cefa1f0440
Signed-off-by: Fred Treven <fred.treven@cirrus.com>