ASoC: cs35l45: Remove limits on compressed stream fragments and fragment size
The limits MIN_FRAGMENTS and MIN_FRAGMENT_SIZE_WORDS were set to
prevent the driver from reading small chunks of data frequently in
order to reduce overhead. Since the data rate of DSP trace logging
is considered not high, the overhead should not be too important and
it is safe to remove these limits.

Also, the size of the DSP trace logging buffer is fixed and is not
massive, therefore, MAX_FRAGMENTS and MAX_FRAGMENT_SIZE_WORDS are
not necessary.

Change-Id: I386d257b84c449a00a0ee0c05e7d4562974e7ff2
Signed-off-by: Qi Zhou <qizhou@opensource.cirrus.com>